requirements.txt
requests
supabase
python-dotenv
flask